backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.9
aiosignal==1.4.0
aiosqlite==0.22.1
alembic==1.13.3
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
asyncpg==0.31.0
attrs==25.4.0
babel==2.17.0
banks==2.3.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
brotli==1.2.0
celery==5.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dirtyjson==1.0.8
distro==1.9.0
ecdsa==0.19.1
email-validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.3
filetype==1.2.0
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.1.0
gradio==6.3.0
gradio_client==2.0.3
greenlet==3.3.0
griffe==1.15.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.2
httpx-sse==0.4.3
huggingface_hub==1.3.2
idna==3.11
importlib_resources==6.5.2
iniconfig==2.3.0
ipykernel==6.29.5
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kombu==5.6.2
langchain==1.2.6
langchain-core==1.2.7
langchain-openai==1.1.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.4
lark==1.3.1
llama-cloud==0.1.46
llama-cloud-services==0.6.91
llama-index-core==0.14.13
llama-index-instrumentation==0.4.2
llama-index-workflows==2.12.2
llama-parse==0.6.91
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.12.4
mdurl==0.1.2
azure-storage-blob>=12.19.0
mistune==3.2.0
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
notebook==7.5.2
notebook_shim==0.2.4
numpy==1.26.4
openai==2.15.0
openai-agents==0.6.6
openpyxl==3.1.5
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pdf2image==1.17.0
pillow==10.4.0
platformdirs==4.5.1
pluggy==1.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.7
pyparsing==3.3.1
PyPDF2==3.0.1
python-docx>=1.1.0
pytesseract==0.3.13
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-json-logger==4.0.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==5.1.1
referencing==0.37.0
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
ruff==0.14.11
safehttpx==0.1.7
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==2.1.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
SQLAlchemy==2.0.35
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
tenacity==9.1.2
terminado==0.18.1
tiktoken==0.12.0
tinycss2==1.4.0
tomlkit==0.12.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.21.1
typer-slim==0.21.1
types-requests==2.32.4.20260107
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.30.6
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==12.0
widgetsnbextension==4.0.15
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
